+++ src/enums/series.py
@@ -1,0 +1,8 @@
+from src.enums import CaseInsensitiveStrEnum
+
+
+class EpisodeFormat(CaseInsensitiveStrEnum):
+    STANDARD = "Standard"
+    DAILY_DATE = "Daily / Date"
+    ANIME_ABSOLUTE = "Anime / Absolute"
+    DVD = "DVD"