+++ CHANGELOG.md
@@ -40,6 +40,7 @@
   - Pops up to preview the renamed folder(s) and file(s) showing a **diff** in red/green colors.
   - Gives the user a chance to confirm the rename or go back and make changes.
 - ImageViewer widget now has a draggable splitter, this way the user can resize/collapse the log/image portion of the widget.
+- Series support during flow to handle mapping/renaming
 
 ### Changed